Add list of transactions to property show
