+++ src/store/modules/property.js
@@ -15,6 +15,7 @@ export default {
     kind: null,
     amount: null,
     currency: null,
+    transactions: [],
     isLoading: true,
     isLoaded: false
   },
@@ -28,6 +29,10 @@ export default {
       commit('START_LOADING', id);
       const item = await api.property(token, { id });
       commit('FINISH_LOADING', item);
+    },
+    async destroyTransaction({ commit }, { token, id }) {
+      await api.destroyTransaction(token, id);
+      commit('FINISH_TRANSACTION_DESTROYING', id);
     }
   },
 
@@ -36,13 +41,17 @@ export default {
       state.id = id;
       state.isLoading = true;
     },
-    FINISH_LOADING(state, { name, kind, amount, currency }) {
+    FINISH_LOADING(state, { name, kind, amount, currency, transactions }) {
       state.name = name;
       state.kind = kind;
       state.amount = amount;
       state.currency = currency.name;
+      state.transactions = transactions;
       state.isLoading = false;
       state.isLoaded = true;
+    },
+    FINISH_TRANSACTION_DESTROYING(state, id) {
+      state.transactions = state.transactions.filter(v => v.id !== id);
     }
   }
 };
